refactored updates
