+++ src/base64/decompressFromBase64URL.ts
@@ -1,15 +1,12 @@
-import { _compress } from "../_compress";
+/*
+ * SPDX-FileCopyrightText: 2013 Pieroxy <pieroxy@pieroxy.net>
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
 import { _decompress } from "../_decompress";
 import { getBaseValue } from "../getBaseValue";
-
-const keyStrBase64URL =  "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_";
-
-export function compressToBase64URL(input: string): string {
-    if (!input) {
-        return "";
-    }
-    return _compress(input, 6, (a) => keyStrBase64URL.charAt(a));
-}
+import { keyStrBase64URL } from "./keyStrBase64URL";
 
 export function decompressFromBase64URL(input: string): string {
     if (!input) {
