+++ src/base64/index.ts
@@ -5,6 +5,8 @@
  */
 
 export { compressToBase64 } from "./compressToBase64";
+export { compressToBetterBase64 } from "./compressToBetterBase64";
+export { compressToBase64URL } from "./compressToBase64URL";
 export { decompressFromBase64 } from "./decompressFromBase64";
-export { compressToBetterBase64, decompressFromBetterBase64 } from "./betterBase64";
-export { compressToBase64URL, decompressFromBase64URL } from "./base64URL";
+export { decompressFromBetterBase64 } from "./decompressFromBetterBase64";
+export { decompressFromBase64URL } from "./decompressFromBase64URL";
