+++ testdata/hello_world/js/base64url.bin
@@ -1,0 +1,1 @@
+BIUwNmD2AEDukCcwBMg
