+++ testdata/hello_world/js/betterbase64.bin
@@ -1,0 +1,1 @@
+BIUwNmD2AEDukCcwBMg==
