+++ testdata/pi/js/base64url.bin
@@ -1,0 +1,1 @@
+MwOgjALGCsCcBMA2axoA5YHZbHsNEiSEw-u20ADPGgWFmIjrDptGJevJVhLBBDh5KmNIwiUiaIpQyxpNSsAhp4KCPDBhMk7KshpZyMLlXBJiCDpwFCPaFS55ta1P1lgaKrdsHU61Dr0-NCaHA5wloTw8PwY0Hzm-PRRNGI8bIjIOMqEngSZKnjAooiiwIyhmJ6U9DyQ0JZojUx4dsrSHIIkSE2apDiIIkjUfJCemCXwgmg6FpIVDrS-y6E83LBSsWoSsNowwJaHoghwSkO1JuYonCqqlITI_XSEsOz8MFpD_KScmPA6Q7QNhvPZvcwIMCdbTJHBILTYCqcLTNFQiAREEosIiYPiNSbsWjAqrTDCeAE2KHmLT8PaxNCkMqM6YSZBZEhGagmWmEYBvGJWEp1HSaPTfEgizA6FCQ6qYIg4bRlUL4Qi4sQEMpZdzsXCUfWNMQAhpIVBDVRS5pSkwaNiWzb1SaIIWif6QRC2Hi1AXgjnuhJvZpCzgiBAUj2nDB4BLcehvRBaTSxLkO6QSXJQmDyEwKqVWaoQh2EnRum0scvyKZYQx8N7hLAVSY0Rg8Z1QuBUFlwQTqJiyczENQxcyqFC5D1UVDwxVQsTVWr6pQmcqzUK0ZCr0jcBliDSSfZS7WQSZKOCj5SwB40NgJD2TLQdqTO2jRkqr1jvYGiduWzi4fYxB6FgmPq9AArMWQ-KC6ziLG8g2PYaj_AuyAqOgzrgdIcCHAcyQWpsZhKLU_KntSzSxBGbxSocMRwB8-B0nwbpSmuyjYCYnwOOqwLTNMCQwJIoJUKo4IVIwnT_KISwYCIVCVnOt4lIQiSKqcVBDBI6xSrUZRaHy-ryHc0zes0DIJkMZolCIWQLlYHCZvwsQRJs9AYGqZQMrifACKOKDYHm_wenZDwPGwyiWDINZZAQaiXkwYyxKQObaTQ8EwJooWhIw3CNJsQzGegShiFQwTsJ4Y6qByp7oPKEjTLimyCqaexQCwAIjBoTBhtgHnRlYYjOhcOBQG6lieHGfHSIsRE0AmIm4hEOhlHwO6ON56Cma0XW5VpGX-XJA5vCgYTILk8jIQCDgOngTQWAGeyuRqbFYGFLSHPI6JIDGVQKA4mZ0HaaqpRkUqSQh-oDVYkigQmgjTGCor_m6-bRrUUACPQBITt4CY4TABJqFlpBWMCUAEPwQUhMUeCXpMHwkMTCSHtM4lEISla5WhjQbJCYaaDF_wONQ30ehBcqXt5QyVOFQErQo_zkHspSyMCBBkPhZRYNFDRQAZKC0G-g32JIhiSBoajlTlCJ8vSKiplAXjrkg_CNFerVPM01AxBwkFQs6ugCOBRDcHy9DraFnhCobDrxocEr0C4sxqrVyB2nyRBQAxRjoLglh0egVwS6IgRnDwFPApYIjxJ48UenyDKhB21ijqoSA5QKbkiIF1T8KISAEKoUtWF1XrELUhx8gISkONcJsh3wbN8b4zCbOwJCQPS9RxUIB3_JoTARZxzc50n9aiM0oJSrEbArTgOnaDKXgeiYg_SDuEa2riMB0e2O98eCoJ14PBYILaYQECBnkknHV46AVCnTMk6GIbdviMXNMiaY-A1YS3YCyHMAZ5TgVDNaXKkA3AvVAhRFsbkoDxTfhEAerJUDKBiMMWiqgB5pmKhA_KnspZBQBJPUISBcRanPvyIm5EdgsFvFkSSZh5B52yAmRgqA7iYj-AtSwH1UDYWdIIZ-uJcBQHKgJeUgkrAuw4KvGOdUhG0H8hwDAFR9FMBzNIPeSdO6u3spOEcCZsCzUMAgWQuIzFK0qjYdAl4MiQEvJsEE5FcbIHwRqRU18B77xEvBFQe8Nw1itI0CQG8dD5giCYPAPV0bsBVCItoccYymQEBY0IG5erhC0JnJmqA6Yqj0raRgCYuRuCYCgTEkAND8hUIkaoswtBowrgkU8l4KTDXTLibgIgkSqxlJITQ8kvLFwKAPAiW5xnAhfned4JB1Icn5oICoewHgDw4H-Ja7pDDLQRt-B6WAV5rFhikPR14EgCDNtPSOMywLaHLAyR5WZ6CNEzGqNeztI68E7rTP23lnAuXxC7bc9h_SdiHBUCyyw74bx4CUIgy5miPmfKwV0_YlBxV-dwaiQN-ZqDZtiBMJUNACXMEBVU-ZlABm4CMfijw_5o14BoT2TZCRZAuP0DBEhyUiC6FNRs4UF5rx2cCTY_dEUGmEj7DcPUWADW6YKZ8EhG7VEEI81KS10aXBidwKQDx9J8hBNUso1Qvz6j3KQUoLBDj6gVpwGqbZ5RuAFbMCoFs7I5VwAgMQexB5-3Ku4_8hrXLow4bEcQUNvjEgBIJHiGwH66UYHvR52Y-x7DXLUAkugJhoQZEg2ZDcNhMNXoLca_Z826iYC-eoTSMq6HYEocK1QODOJrvOnMLBs7NBUZWASmMdLVK-Oux11KK5UmqY0YS5My4YAvvIPAiaew0CSs6IWtV2JBiibgeEqIMAYXKTkWQ8hgRm24Mi2KIh6gJF3JUFa3ZzJCkMBW7MLAORtR8vEO0-RgZCFwFub4_wQE1FsKQYEudwM2BpjSXAcARGHCZJwMyDwN5iEwUkEOESlBsEvaECUGMRR7HIGyKEVKOWgvMCWV4YQ4IgRDkkQwmMaX2ReQIBAcJPjmg0E0i-zp5RjoYnYoK5RI4OC-voZxeJ5RWjbgzDGTiWUeksBZ5QHBqKAWrP4TZUJm0Y3sg8QgZghzVBkOEL6MRSwDJs1kLiEIdjzujVJalMt7inSsV-lpmaZCFPlGkgD6EzQBbpekYYgxJCQKaMo9iDwxx3EnHkPBZnlRGP4DQ4VRrU7VmtHS1ThnV4O3C-qm4sNcjr02hltgDRIKslNnMCpsx5RwBtEU6oAJvIlJysHa4DEXa40gJqiYi4JBSUFLncNl5dwyOUhtIhFbB51wWD0UyUltOcGjIYJxQriS4l0G8QQ_mShsB2BKjpR15QwF3PqNQ3pg4X0zlySwga5IBMw2XUMqFJhA8JkqfEupiTaBEsq7oUolZ-ETNiAwl5zC8kQwWCie1lDepG6h4ZDQZLBAwAWHw2ArzVhoOlycIdXKuLXv4HQz26JPIXB8TQVhvaNl7jDq-wd0Y7FRIs055ygz7nNssEo6AqAcmdEjQk18FGrblJoWbQxrTaiizkhi0hUI8hKv6VkXhGTOlYNE31W18QS-fH1p0rZznVsMHGrt_mxo4juLQSN0hTj5WWV4N0IVGxhGkLIIDdRg0xDyvsFIP3IShqx7EjstQWjALbAoXuX1bUaBVB4bSxxqlxTbajf4Dd7VlET-465OYkH1UDWSHX4IeC2DtqQCEeiQlB6YBRo7xd-LDXGSjryUJOpJoUbIC965T651CqbkZ2V07oS-wOFRWYM0PPYEyfWnheB1tw_1MovFtiEASNR02H0tTppvYoSilcnkYOV0EDQs4qvCzu4AgD0G-JsLTIJPjubDRMJN7KLFRGQGKMIMXoyjDpUOSE4qrFLBtIwA_KrNfFOMTHYPKp8G3MhCUMoM4oli9HmC2A8nFMkmTHGmyjkBjPyDAJnGTIhqiAIJMPopMAyNlLWA5rxkikEgCKDAulNNNjQdoMhDfKevciwqDDBHBMiDZEajWHEqaFLFgGeOwgQVuKSlkECLNMEL0NqkMDiGaHGqeAIQQoYKgOVl3ArA4MFNrp9rMIEMdqQtzEUoIPsB8mAiQEQJrPohMCgJmDTuUEkPARoOQLCopmFJVFMLRkwh0taBNLYckMTAGneFAZRkRJWKHCJvOOdCFBfFcCIujPKsFoFOFN2CUgtLMAUK6HBNIelEHo2JrF3JqjxHNoKNDABpsheBglPnpA3MAqSkmgQFeDgmBKetCh8h8CYC2hxFDCBP1OIG-OuuRHoUrrGJULCuIWyP0GMCQFAI-GoFhFhFlilqXmsP-i4SEKwp1EUMwC3rAvpv7F-CxLYbQBBBiKSGAlaAgBZOCFUIIS4aLsmLRmYuxhbKFFDHsEcC5DRFqN5jHotkBtuHqM0L7HFDsC_lMBxCgB2BAibFfHYMmAQRBJrJxvuECkIl-AihMtlCwMEGbDPPji1GxjTrdr9lILQJsHyIGNgCTLhr4HZo-gtI3HtnEg0Bscuu9LrD8nyBRgnNZNQKnmsB-noDEfYtKDFFgLFLfoxrEiPF6FCowAwCjuBMPAUP3Kpu7j9qMLxOibqAaAaOJFhnHK4TptoNQPBDbMUMEXvOPEaGscgJcKIKCknO8FLEmFzuGv5s4NSPOupspF6FnFEJ9qCvuAorTOBhftQBhGwI9pWGMEnFXNFLIE9CEBwDEAyGKirE8uaGDlqUMv-AwH9GrAsAkELNZB6DQLEIkPRj9skhSL6uwDfLjAApqG2YMGQKicoryvrAXPrHuIaBVDSu4iSjTveiEomKYOBJCGDtaG3BTO-rHPBJejgFwF8JohWi2cehFCwmWLpISBmhkNJjSKbExHETTIMocKiBYOvMwHDpeHUF1K7oZvcgaH4IVBOCVPGCeJnnstyAMO2MKopiyvcB2DOl7I-kwgmMEq1L3D2CEJSNQDQUdIppRuUPjvResgQZJGzjdNcaopPH3PzO2LNMZOuppMqt5luBtESQotKOvEzIJKnFYDMELKYRGdTsXk6vapUH5HFLbgIqGtIuujAGzAmXeS-I6Q8pemhjcVRGIDSI-m2QaO4ImibONDwrVDoctGQXmFolmG8LcA2OstzH2F0KTpaHzg3JSM_l_P3MAtyjkv-oivcgrPfpMEkBLKLGQOEc0dpStGNAgNKM-Lho2DktTP2GYmeaGQ9McDYDiH-iQLMOGgFgqNUpnEKGplGAyKwuDrBSTI8AUKdEerLliMBsGqWbDj9CkKMqqjbJXOZOPPZd_LTObHqbzrMNIUMDCXsqarDDFGUg1a4EoF3vXIHMdG0PjJiDhIBupPpExFQNnDiFYpSWCM_izmvBxIda3BlkKvVNchPNyuiVXNwd-LZWuOEUyW8o4nyklDrnDB2nTHxIpt0DdOwinrjOqoWppvKHMC5OyhQVcd6oEK4CnrVGYHoqQuFNGO5naOyFlGNNSpakuHiMRI4sSIZEkWCOXrBMEIhrTDdG0BFLYNbO4MRpmhIFPLPlcAxHDKTDTH5TjrNq2cJPQqboLEuJqkhODtdsZNsOIl-VCN3MqMHPRjGFocviolMpkBfjCl8OMAuJVebu_HoQDQOLQJpKivRkqWkDSlQAmUvO3sdnMIJJ_jUDZGTLBkVjog_o3BRnVpYcWm7NNnNHcdfMYIameFDGaEggqB0BcA5jVJnoYOoPBHGYBMWuqOfBKoIRsD3vcvXEFcPK4nwNGlINXiytEdaP4HVNgO_oNFwOBlGE0HBsOZBK2hRCKOyNZFdBzmXPosVEVl9GtjMhOTtoyIqG-POSnhLKBrbtmavDnDxcPDkFaHONMXnDVMcEaNSDMGZgtjZB0LxnEtzAZMwJdUiMWuZD0YxmrOzqukkb4Kla7miF9thLEs4tWXGtKd4KVP7XMsuCNgnuvosGcKSDcMwcEvTWBFJpnjwJZlqCJd7E4ViFqfA4dWDXGF0LOiNtbEzAnHSg8AWRWUCDggNK2ZVNEKzSor8KHKwvrEpFyltgbMJKnKHMQq2CjaEoPLlKCFRARZBOxhmFrsOFgGIzTqbJSenUrPxMIbOBKQ1kwOqFqF9BflwDNgPKgK7TEZ_IaOzqGYSiKL2gwHovokmI1N6r0kqPqOFvZCzBstkDeOugCP4qZPgNsPRuioHBGhNjyV9nNHCBMdsD1OhauPrUmN3D7dmITGiVEuVo0N0peRlKPo8LwG1JZMLEtnBk3R0gikmCE8VBCB0qGJsinhbEIusgZQJv8HZDygMkrpLWikA73KGDSAcMIjQf4LgingroNHyOBrtmiBWSUkakUK2umo1FpoYSISHlsRbDxq0B0F-oE0FOmHgCMPlZxg6Hk41K0KerMc_HRu7jeusMNFSSzpWkCBeKyIddDqhBKEsNEvoqYK3GtI3IUpwFEJVKnfrTQykHciQCAa4FrvVVJTg3pEuHJC0tCm-KDjfuyl4xYpPCBvSMURTNIqeqjn7IivEqGgoPAQmEOCotuE-HzAuotmKfzKhHHHBmPUHn8PxmvLyEBq0OzmOTUMIZ5N-EkZnaTX-hhOvmOQKEBOSEUKVjMlEs-Ftpmj6kyA_iwsCXmsuA4LnWDCkcGulMo4Lp1VptYE0pouPD2GNILBKF6Po8wHS0hBvNkj-dcgmXeOuOMjMJHA0E-T3OTLRPHi2APt7LxNsQ7MeIXYIX9CGOvnUNHhcwOBA0yukApITjyY6uSAMEYK7mQDMxOUhNHiGLVEiAsA3K5DQU4tZZtqiGxLeHiLlN-EVIlDEhFGvKqKBqBn7sYhUmQmrHDBbIMj7YEGC9GL0N6CTSoCOHtsop1SBFUHCLcItk-MrhZMtMdFhOEEtDVEZP9cVIiB5FrPo8tcSJwCcomfo64bGLrJ6h4Z7Bsp02yRhMYA5tGuAcRIEAKhnFgltsO9CYInEqCiUq4tWrsPgHMDXPojDBUvKnBlYoE19oqEVgEk5Oc_YpdJqN400O9FamGLlJEKsMEq7AuFpdylrmUEAA
