+++ testdata/repeated/js/betterbase64.bin
@@ -1,0 +1,1 @@
+IY18ZXTt/DFQEarejmvZ7v+DCjsBjM8iyq6m2u+hxp5mgE3Y86+597/4GChwkbyA==
