--- tests/suite/compressTattoo.ts
@@ -1,7 +1,0 @@
-import { expect, test } from "vitest";
-
-export function compressKnownString_test(compressFunc, data, compressedData) {
-    test(`expected compression result`, () => {
-        expect(compressFunc(data)).toEqual(compressedData);
-    });
-}
