--- tests/suite/index.ts
@@ -1,12 +1,0 @@
-export * from "./helloWorld";
-export * from "./nullResponse";
-export * from "./emptyResponse";
-export * from "./undefinedResponse";
-export * from "./utf16Response";
-export * from "./repeatingString";
-export * from "./randomString";
-export * from "./longString";
-export * from "./urlSafe";
-export * from "./inchangableChars";
-export * from "./compressTattoo";
-export * from "./decompressTattoo";
